In addition to implementing almost all of constructor, I went back and made other minor changes:
-Classroom now has a getStudentsIterator() function
-Persistent has a getFile_name() function, and tracks file_path.
PracticeItem now has a constructor that will create a PracticeItem
from a song.
-Practice list has getters & setters for due day / assigned day. Also
added function to check if list is empty. Created a toString function.
-ProgressReporter now takes an instructor as a constructor argument,
and no longer works as a singleton. Every call to report(String r) will
include a newline after the entry.
-Song now has a toString() function.

Also to note, I made necassary changed in MusicSchool to take into account
that Instructor's Constructor must include the songDB for the program.